OpenGL renderer: Build without "OPENGLRENDERER_NO_STATE_CLEANUP" and resolved "warning C4365: 'argument': conversion from 'GLint' to 'GLenum', signed/unsigned mismatch" etc. compiler warnings (there's no "glGetIntegerv()" for unsigned integer)

+++ Renderer/RendererBackend/OpenGLRenderer/src/Buffer/IndexBufferBind.cpp
@@ -52,7 +52,7 @@ namespace OpenGLRenderer
 
 		#ifndef OPENGLRENDERER_NO_STATE_CLEANUP
 			// Be polite and restore the previous bound OpenGL element array buffer
-			glBindBufferARB(GL_ELEMENT_ARRAY_BUFFER_ARB, openGLElementArrayBufferBackup);
+			glBindBufferARB(GL_ELEMENT_ARRAY_BUFFER_ARB, static_cast<GLuint>(openGLElementArrayBufferBackup));
 		#endif
 	}
