+++ Renderer/RendererBackend/OpenGLRenderer/src/Buffer/TextureBufferBind.cpp
@@ -54,7 +54,7 @@ namespace OpenGLRenderer
 
 			#ifndef OPENGLRENDERER_NO_STATE_CLEANUP
 				// Be polite and restore the previous bound OpenGL texture buffer
-				glBindBufferARB(GL_TEXTURE_BUFFER_ARB, openGLTextureBufferBackup);
+				glBindBufferARB(GL_TEXTURE_BUFFER_ARB, static_cast<GLuint>(openGLTextureBufferBackup));
 			#endif
 		}
 
@@ -73,7 +73,7 @@ namespace OpenGLRenderer
 
 			#ifndef OPENGLRENDERER_NO_STATE_CLEANUP
 				// Be polite and restore the previous bound OpenGL texture
-				glBindTexture(GL_TEXTURE_BUFFER_ARB, openGLTextureBackup);
+				glBindTexture(GL_TEXTURE_BUFFER_ARB, static_cast<GLuint>(openGLTextureBackup));
 			#endif
 		}
 	}
@@ -101,7 +101,7 @@ namespace OpenGLRenderer
 
 		#ifndef OPENGLRENDERER_NO_STATE_CLEANUP
 			// Be polite and restore the previous bound OpenGL texture buffer
-			glBindBufferARB(GL_TEXTURE_BUFFER_ARB, openGLTextureBufferBackup);
+			glBindBufferARB(GL_TEXTURE_BUFFER_ARB, static_cast<GLuint>(openGLTextureBufferBackup));
 		#endif
 	}
