+++ Renderer/RendererBackend/OpenGLRenderer/src/Buffer/VertexArrayVaoBind.cpp
@@ -100,13 +100,13 @@ namespace OpenGLRenderer
 
 		#ifndef OPENGLRENDERER_NO_STATE_CLEANUP
 			// Be polite and restore the previous bound OpenGL vertex array
-			glBindVertexArray(openGLVertexArrayBackup);
+			glBindVertexArray(static_cast<GLuint>(openGLVertexArrayBackup));
 
 			// Be polite and restore the previous bound OpenGL element array buffer
-			glBindBufferARB(GL_ELEMENT_ARRAY_BUFFER_ARB, openGLElementArrayBufferBackup);
+			glBindBufferARB(GL_ELEMENT_ARRAY_BUFFER_ARB, static_cast<GLuint>(openGLElementArrayBufferBackup));
 
 			// Be polite and restore the previous bound OpenGL array buffer
-			glBindBufferARB(GL_ARRAY_BUFFER_ARB, openGLArrayBufferBackup);
+			glBindBufferARB(GL_ARRAY_BUFFER_ARB, static_cast<GLuint>(openGLArrayBufferBackup));
 		#endif
 	}
